Summarize the Dataset
1. Dimensions of Dataset
how many instances (rows) and how many attributes (columns)
You should see 150 instances and 5 attributes: (150, 5)
2. Peek at the Data
You should see the first 20 rows of the data
3. Statistical Summary
This includes the count, mean, the min and max values as well as some percentiles.
We can see that all of the numerical values have the same scale (centimeters) and similar ranges between 0 and 8 centimeters.
4. Class Distribution
Let’s now take a look at the number of instances (rows) that belong to each class. We can view this as an absolute count.