2D Slicing: Split Input and Output Features
For the input features, we can select all rows and all columns except the last one by specifying ‘:’ for in the rows index, and :-1 in the columns index.
X = [:, :-1]
For the output column, we can select all rows again using ‘:’ and index just the last column by specifying the -1 index.
y = [:, -1]
Running the example prints the separated X and y elements. Note that X is a 2D array and y is a 1D array.